+++ storage/books.json
@@ -1,1 +1,1 @@
-[{"title":"book 1","author":"author 1"},{"title":"book 2","author":"tolu"},{"title":"book 3","author":"author 3"},{"title":"book 4","author":"author 4"},{"title":"book 5","author":"author 5"}]
+[{"title":"Fire","author":"Martin"}]
