+++ storage/people.json
@@ -1,1 +1,1 @@
-[{"type":"Teacher","name":"olag","age":"56","specialization":"marine","parent_permission":"marine"}]
+[]
