+++ storage/rentals.json
@@ -1,1 +1,1 @@
-[{"date":"20-04-2021","book":{"title":"book 1","author":"author 1"},"person":{"type":"Teacher","name":"olag","age":"56","parent_permission":"marine","rentals":[{"date":"20-04-2021"}]}}]
+[]
